Implemented Google Auth!
Had to follow a few guides and look at some documentation, but I think that I did it! I still need to implement the different button states.

Also, I'm wondering if we want to store the google profile pictures, which could be a feature that we add.  in a different database migration