Better Google Sign in Button
Found a better version of the Google sign in button that uses bootstrap styling to make it look like a proper button. This improved version includes hover and active animations courtesy of bootstrap.

Also added in a temporary logo. I don't really like how it appears when reduced in size, so I might end up trying something else later on.